+++ src/main/kotlin/com/github/victorrentea/livecoding/complexity/ComplexityRenderer.kt
@@ -1,4 +1,4 @@
-package com.github.victorrentea.livecoding.extracthints
+package com.github.victorrentea.livecoding.complexity
 
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.editor.markup.CustomHighlighterRenderer
@@ -9,20 +9,20 @@ import java.awt.FontMetrics
 import java.awt.Graphics
 
 
-class MethodComplexityRenderer(private val complexity: Int) : CustomHighlighterRenderer {
+class ComplexityRenderer(private val complexity: Int) : CustomHighlighterRenderer {
     override fun paint(editor: Editor, highlighter: RangeHighlighter, g: Graphics) {
         g.color = when (complexity) {
 //            in 2..4 -> Color.green
-            in 5..8 -> Color(255,153,51)
-            in 9..1000 -> Color.red
+            in 5..9 -> Color(255,153,51)
+            in 10..1000 -> Color.red
             else -> return
         }
         val methodSignatureEndPoint =
             editor.logicalPositionToXY(editor.offsetToLogicalPosition(highlighter.endOffset))
 
-        val RIGHT_MARGIN_FROM_TEXT = scale(10)
+        val RIGHT_MARGIN_FROM_TEXT = scale(8)
 
-        val s = "Cognitive Complexity : $complexity"
+        val s = " $complexity = Cognitive Complexity"
 
         val x = methodSignatureEndPoint.x + RIGHT_MARGIN_FROM_TEXT
         val y = methodSignatureEndPoint.y + editor.lineHeight / 2